FEAT: 토픽으로 게시글 조회

+++ Odya-iOS/Core/Community/CommunityCompose/CommunityComposeViewModel.swift
@@ -27,7 +27,7 @@ final class CommunityComposeViewModel: ObservableObject {
     @Published var isSubmitted: Bool = false
     
     // MARK: Create
-    func createCommunity(content: String, visibility: String = "PUBLIC", placeId: String?, travelJournalId: Int64?, topicId: Int64?, imageData: [ImageData]) {
+    func createCommunity(content: String, visibility: String = "PUBLIC", placeId: String?, travelJournalId: Int?, topicId: Int?, imageData: [ImageData]) {
         let uiImageList = imageData.map { $0.image }
         _Concurrency.Task {
             let _ = await webpConverter.processImages(uiImages: uiImageList)
